[ntuple] Move `SerializeXxx` routines for fundamental types to a named namespace
This commit moves the `SerializeXxx` routines for fundamental integer types
to a named namespace. The aim is to make these functions reusable by other
parts of the RNTuple code, e.g., alternative backends, etc.
